+++ android/KMEA/app/src/main/java/com/tavultesoft/kmea/KMManager.java
@@ -114,6 +114,12 @@ public final class KMManager {
     TABLET;
   }
 
+  public enum Tier {
+    ALPHA,
+    BETA,
+    STABLE
+  }
+
   private static InputMethodService IMService;
   private static boolean debugMode = false;
   private static boolean shouldAllowSetKeyboard = true;
@@ -214,7 +220,6 @@ public final class KMManager {
   public static final String KMDefault_KeyboardName = "EuroLatin (SIL) Keyboard";
   public static final String KMDefault_LanguageName = "English";
   public static final String KMDefault_KeyboardFont = "DejaVuSans.ttf";
-  public static final String KMDefault_KeyboardVersion = "1.8.1";
   public static final String KMDefault_KeyboardKMP = KMDefault_PackageID + FileUtils.KEYMANPACKAGE;
 
   // Default Dictionary Info
@@ -256,8 +261,37 @@ public final class KMManager {
     return getResourceRoot() + KMDefault_UndefinedPackageID + File.separator;
   }
 
+  public static FormFactor getFormFactor() {
+    String device_type = appContext.getResources().getString(R.string.device_type);
+
+    return device_type.equals("AndroidMobile") ? FormFactor.PHONE : FormFactor.TABLET;
+  }
+
   /**
-   * Extract app version #.#.# from VERSION_NAME
+   * Extract KMEA tier from versionName. Uses parameter so we can unit test.
+   * @param versionName String - If not provided, determine tier from
+   *                    com.tavultesoft.kmea.BuildConfig.VERSION_NAME
+   * @return Tier (ALPHA, BETA, STABLE)
+   */
+  public static Tier getTier(String versionName) {
+    if (versionName == null || versionName.isEmpty()) {
+      versionName = com.tavultesoft.kmea.BuildConfig.VERSION_NAME;
+    }
+    Pattern pattern = Pattern.compile("^(\\d+\\.\\d+\\.\\d+)-(alpha|beta|stable)-.*");
+    Matcher matcher = pattern.matcher(versionName);
+    if (matcher.matches() && matcher.groupCount() >= 2) {
+      switch (matcher.group(2)) {
+        case "alpha": return Tier.ALPHA;
+        case "beta": return Tier.BETA;
+        default:
+          return Tier.STABLE;
+      }
+    }
+    return Tier.STABLE;
+  }
+
+  /**
+   * Extract KMEA version #.#.# from VERSION_NAME
    * @return String
    */
   public static String getVersion() {
@@ -721,7 +755,7 @@ public final class KMManager {
    * 2. Undefined package ID set to "cloud"
    * 3. Undefined version set to the latest file version
    * 4. Check sil_euro_latin only added once
-   * Using the old keyboards list as paramter instead of keyboards file so this can be
+   * Using the old keyboards list as parameter instead of keyboards file so this can be
    * unit test.
    * @param context Context
    * @param dat_list ArrayList<HashMap<String, String>> Old keyboards list
@@ -767,7 +801,7 @@ public final class KMManager {
       if (keyboardID.equals("us") || keyboardID.equals("european") || keyboardID.equals("european2") ||
           keyboardID.equals("sil_euro_latin")) {
         if (!defaultKeyboardInstalled) {
-          list.add(Keyboard.DEFAULT_KEYBOARD);
+          list.add(Keyboard.getDefaultKeyboard(context));
           defaultKeyboardInstalled = true;
         }
       } else {
@@ -1348,7 +1382,10 @@ public final class KMManager {
       try {
         File kmpJSONFile = new File(path);
         if (!kmpJSONFile.exists()) {
-          return null;
+          if (!KMManager.isTestMode()) {
+            KMLog.LogError(TAG, path + " not found. Returning version 1.0");
+          }
+          return "1.0";
         }
         JSONParser jsonParser = new JSONParser();
         JSONObject kmpObject = jsonParser.getJSONObjectFromFile(kmpJSONFile);
@@ -1602,12 +1639,6 @@ public final class KMManager {
     return result;
   }
 
-  public static FormFactor getFormFactor() {
-    String device_type = appContext.getResources().getString(R.string.device_type);
-
-    return device_type.equals("AndroidMobile") ? FormFactor.PHONE : FormFactor.TABLET;
-  }
-
   public static boolean isHelpBubbleEnabled() {
     boolean retVal = true;
 
@@ -1707,8 +1738,10 @@ public final class KMManager {
             langId = keyboardInfo.getLanguageID();
             InAppKeyboard.setKeyboard(keyboardInfo);
           } else {
-            langId = KMDefault_LanguageID;
-            InAppKeyboard.setKeyboard(Keyboard.DEFAULT_KEYBOARD);
+            // Revert to default (index 0)
+            keyboardInfo = KMManager.getKeyboardInfo(context, 0);
+            langId = keyboardInfo.getLanguageID();
+            InAppKeyboard.setKeyboard(keyboardInfo);
           }
 
           registerAssociatedLexicalModel(langId);
@@ -1928,8 +1961,10 @@ public final class KMManager {
             langId  = keyboardInfo.getLanguageID();
             SystemKeyboard.setKeyboard(keyboardInfo);
           } else {
-            langId = KMDefault_LanguageID;
-            SystemKeyboard.setKeyboard(Keyboard.DEFAULT_KEYBOARD);
+            // Revert to default (index 0)
+            keyboardInfo = KMManager.getKeyboardInfo(context, 0);
+            langId = keyboardInfo.getLanguageID();
+            SystemKeyboard.setKeyboard(keyboardInfo);
           }
 
           registerAssociatedLexicalModel(langId);