+++ android/KMEA/app/src/main/java/com/tavultesoft/kmea/data/KeyboardController.java
@@ -60,20 +60,21 @@ public class KeyboardController {
       list = new ArrayList<Keyboard>();
       if (keyboards_dat.exists() && !keyboards_json.exists()) {
         try {
-          // Migrate installed_keyboards.dat to installed_keyboards.json
+          // Migrate installed_keyboards.dat to keyboards_list.json
           ObjectInputStream inputStream = new ObjectInputStream(new FileInputStream(keyboards_dat));
           ArrayList<HashMap<String, String>> dat_list = (ArrayList<HashMap<String, String>>) inputStream.readObject();
           inputStream.close();
           list = KMManager.updateOldKeyboardsList(context, dat_list);
         } catch (Exception e) {
           KMLog.LogException(TAG, "Exception migrating installed_keyboards.dat", e);
-          list.add(Keyboard.DEFAULT_KEYBOARD);
+          list.add(Keyboard.getDefaultKeyboard(context));
         }
       } else if (keyboards_json.exists()) {
+        JSONArray json_list = null;
         try {
-          // Get installed keyboards from installed_keyboards.json
+          // Get installed keyboards from keyboards_list.json
           JSONParser jsonParser = new JSONParser();
-          JSONArray json_list = jsonParser.getJSONObjectFromFile(keyboards_json, JSONArray.class);
+          json_list = jsonParser.getJSONObjectFromFile(keyboards_json, JSONArray.class);
           if (json_list != null) {
             // Can't foreach JSONArray
             for (int i=0; i<json_list.length(); i++) {
@@ -83,15 +84,18 @@ public class KeyboardController {
                 list.add(k);
               }
             }
+          } else {
+            KMLog.LogError(TAG, KMFilename_Installed_KeyboardsList + " is null");
           }
         } catch (Exception e) {
-          KMLog.LogException(TAG, "Exception reading installed_keyboards.json", e);
-          list.add(Keyboard.DEFAULT_KEYBOARD);
+          KMLog.LogExceptionWithData(TAG, "Exception reading " + KMFilename_Installed_KeyboardsList,
+            KMFilename_Installed_KeyboardsList, json_list, e);
+          list.add(Keyboard.getDefaultKeyboard(context));
         }
       } else {
         // No installed keyboards lists
         // 3rd-party OEM may not have sil_euro_latin, so don't assign a default keyboard
-        //list.add(Keyboard.DEFAULT_KEYBOARD);
+        //list.add(Keyboard.getDefaultKeyboard(context));
         Log.w(TAG, "initialize with no default keyboard");
       }